Module epic_util::zip

Compress and decompress zip bz2 archives

Functions

compress

Compress a source directory recursively into a zip file. Permissions are set to 644 by default to avoid any unwanted execution bits.

create_zip

Create a zip archive from source dir and list of relative file paths. Permissions are set to 644 by default.

decompress

Decompress a source file into the provided destination path.

extract_files

Extract a set of files from the provided zip archive.